renaming views folder to "Resources"
