+++ frontstage_api/__init__.py
@@ -16,4 +16,4 @@ logger_initial_config(service_name='ras-frontstage-api', log_level=app.config['L
 
 
 import frontstage_api.error_handlers  # NOQA # pylint: disable=wrong-import-position
-import frontstage_api.views.secure_messaging  # NOQA # pylint: disable=wrong-import-position
+import frontstage_api.Resources.secure_messaging  # NOQA # pylint: disable=wrong-import-position
